+++ src/Quartz.Tests.Integration/RAMJobStoreTest.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 
 using NUnit.Framework;
 
+using Quartz.Impl;
 using Quartz.Impl.Matchers;
 
 namespace Quartz.Tests.Integration
@@ -71,6 +72,7 @@ namespace Quartz.Tests.Integration
         protected abstract Task<IScheduler> CreateScheduler(string name, int threadPoolSize);
 
         [Test]
+        [Category("db-sqlserver")]
         public async Task TestBasicStorageFunctions()
         {
             IScheduler sched = await CreateScheduler("testBasicStorageFunctions", 2);
@@ -359,7 +361,9 @@ namespace Quartz.Tests.Integration
         [Test]
         public async Task TestDurableStorageFunctions()
         {
-            IScheduler sched = await CreateScheduler("testDurableStorageFunctions", 2);
+            const string schedulerName = "testDurableStorageFunctions";
+            SchedulerRepository.Instance.Remove(schedulerName + "Scheduler"); // workaround prior test cleanup - relates to issue in #1453
+            IScheduler sched = await CreateScheduler(schedulerName, 2);
             await sched.Clear();
 
             // test basic storage functions of scheduler...