+++ src/Quartz/Impl/JobDetailImpl.cs
@@ -22,7 +22,6 @@
 using System;
 using System.Collections.Concurrent;
 using System.Globalization;
-using System.Reflection;
 
 using Quartz.Util;
 
@@ -50,7 +49,7 @@ namespace Quartz.Impl
         /// </summary>
         /// <param name="jobType">The type for which information will be searched</param>
         /// <returns>
-        /// An <see cref="JobTypeInformation"/> object that describe specified type 
+        /// An <see cref="JobTypeInformation"/> object that describe specified type
         /// </returns>
         public static JobTypeInformation GetOrCreate(Type jobType)
         {
@@ -107,20 +106,6 @@ namespace Quartz.Impl
         private JobKey key = null!;
 
         /// <summary>
-        /// Create a <see cref="IJobDetail" /> with no specified name or group, and
-        /// the default settings of all the other properties.
-        /// <para>
-        /// Note that the <see cref="Name" />,<see cref="Group" /> and
-        /// <see cref="JobType" /> properties must be set before the job can be
-        /// placed into a <see cref="IScheduler" />.
-        /// </para>
-        /// </summary>
-        public JobDetailImpl()
-        {
-            // do nothing...
-        }
-
-        /// <summary>
         /// Create a <see cref="IJobDetail" /> with the given name, default group, and
         /// the default settings of all the other properties.
         /// If <see langword="null" />, SchedulerConstants.DefaultGroup will be used.
@@ -144,7 +129,7 @@ namespace Quartz.Impl
         {
             Name = name;
             Group = group;
-            JobType = jobType;
+            JobType = new JobType(jobType);
         }
 
         /// <summary>
@@ -163,7 +148,7 @@ namespace Quartz.Impl
         {
             Name = name;
             Group = group;
-            JobType = jobType;
+            JobType = new JobType(jobType);
             Durable = isDurable;
             RequestsRecovery = requestsRecovery;
         }
@@ -178,11 +163,11 @@ namespace Quartz.Impl
         /// <param name="isDurable">if set to <c>true</c>, job will be durable.</param>
         /// <param name="requestsRecovery">if set to <c>true</c>, job will request recovery.</param>
         /// <param name="jobDataMap">The data that is associated with the <see cref="IJob" />.</param>
-        /// <param name="disallowConcurrentExecution">Indicates whether or not concurrent exection of the job should be disallowed.</param>
+        /// <param name="disallowConcurrentExecution">Indicates whether or not concurrent execution of the job should be disallowed.</param>
         /// <param name="persistJobDataAfterExecution">Indicates whether or not job data should re-stored when execution of the job completes.</param>
         /// <exception cref="ArgumentNullException"><paramref name="key"/> is <see langword="null"/>.</exception>
         internal JobDetailImpl(JobKey key,
-                               Type? jobType,
+                               JobType jobType,
                                string? description,
                                bool isDurable,
                                bool requestsRecovery,
@@ -191,7 +176,7 @@ namespace Quartz.Impl
                                bool? persistJobDataAfterExecution)
         {
             Key = key;
-            JobType = jobType!;
+            JobType = jobType;
             Description = description;
             Durable = isDurable;
             RequestsRecovery = requestsRecovery;
@@ -303,30 +288,7 @@ namespace Quartz.Impl
             private set => description = value;
         }
 
-        /// <summary>
-        /// Get or sets the instance of <see cref="IJob" /> that will be executed.
-        /// </summary>
-        /// <exception cref="ArgumentException">
-        /// if jobType is null or the class is not a <see cref="IJob" />.
-        /// </exception>
-        public virtual Type JobType
-        {
-            get => jobType;
-            private set
-            {
-                if (value == null)
-                {
-                    throw new ArgumentException("Job class cannot be null.");
-                }
-
-                if (!typeof(IJob).IsAssignableFrom(value))
-                {
-                    throw new ArgumentException("Job class must implement the Job interface.");
-                }
-
-                jobType = value;
-            }
-        }
+        public JobType JobType { get; private set; }
 
         /// <summary>
         /// Get or set the <see cref="JobDataMap" /> that is associated with the <see cref="IJob" />.
@@ -339,6 +301,7 @@ namespace Quartz.Impl
                 {
                     jobDataMap = new JobDataMap();
                 }
+
                 return jobDataMap;
             }
 
@@ -471,7 +434,7 @@ namespace Quartz.Impl
         {
             //doesn't consider job's saved data,
             //durability etc
-            return detail != null && detail.Name == Name && detail.Group == Group && detail.JobType == JobType;
+            return detail != null && detail.Name == Name && detail.Group == Group && detail.JobType.Equals(JobType);
         }
 
         /// <summary>
@@ -516,7 +479,7 @@ namespace Quartz.Impl
 
         public virtual JobBuilder GetJobBuilder()
         {
-            JobBuilder b = JobBuilder.Create()
+            return JobBuilder.Create()
                                      .OfType(JobType)
                                      .RequestRecovery(RequestsRecovery)
                                      .StoreDurably(Durable)
@@ -525,8 +488,6 @@ namespace Quartz.Impl
                                      .PersistJobDataAfterExecution(PersistJobDataAfterExecution)
                                      .WithDescription(description)
                                      .WithIdentity(Key);
-
-            return b;
         }
     }
 }